third_party/blink/web_tests/platform/mac-mac13/external/wpt/pointerevents/pointerevent_to_slotted_target_pen-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer events from pen to slotted element and shadow-host
  assert_equals: events received expected "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@slot,pointerleave@parent,pointerleave@host" but got "pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerover@child,pointerenter@host,pointerenter@parent,pointerenter@slot,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@slot,pointerleave@parent,pointerleave@host"
Harness: the test ran to completion.
